Create new file for Pipeline operators
